# docs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "VA CMS REST API",
    "description": "Headless content management API for VA CMS. BRD FR-DEV-01: OpenAPI 3.0 spec auto-generated from controllers.",
    "contact": {
      "name": "VA CMS Team",
      "email": "cms@va.gov"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/auth/login": {
      "get": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ack",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ack",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter your JWT access token (issued by /api/auth/login).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}